pyproject.toml
[tool.black]
line-length = 90

[tool.poetry]
name = "s-q-r-o-phishing-detector"
version = "0.1.0"
description = ""
authors = ["bnbong <bbbong9@gmail.com>"]
readme = "README.md"
packages = [{include = "s"}]

[tool.poetry.dependencies]
python = "3.12.4"
beautifulsoup4 = "^4.12.3"
html2text = "^2024.2.26"
langdetect = "^1.0.9"
keras-bert = "^0.89.0"
keras = "^3.6.0"
tensorflow = "^2.17.0"
pytest = "^8.3.3"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
fastapi-slim = "^0.115.2"
pydantic-settings = "^2.6.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
uvicorn = "^0.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
